restard, cleaner code
